requirements.txt
discord.py>=2.3.2
yt-dlp>=2024.3.10
apscheduler>=3.10.4
inspirobot>=1.1.3
